refactor: unmarshallers for manifest leaf types
