+++ manifest/unmarshal.go
@@ -24,91 +24,6 @@ import (
 
 // --- Helpers ---
 
-func (f *ClassField) UnmarshalJSON(data []byte) error {
-	type Alias ClassField
-	aux := &struct {
-		Deprecated json.RawMessage `json:"deprecated"`
-		*Alias
-	}{
-		Alias: (*Alias)(f),
-	}
-	if err := json.Unmarshal(data, &aux); err != nil {
-		return err
-	}
-	if len(aux.Deprecated) != 0 && string(aux.Deprecated) != "null" {
-		// Try bool
-		var b bool
-		if err := json.Unmarshal(aux.Deprecated, &b); err == nil {
-			f.Deprecated = DeprecatedFlag(b)
-			return nil
-		}
-		// Try string
-		var s string
-		if err := json.Unmarshal(aux.Deprecated, &s); err == nil {
-			f.Deprecated = DeprecatedReason(s)
-			return nil
-		}
-		// Ignore other types
-	}
-	return nil
-}
-
-func (p *Parameter) UnmarshalJSON(data []byte) error {
-	var proxy map[string]json.RawMessage
-	if err := json.Unmarshal(data, &proxy); err != nil {
-		return err
-	}
-	var depRaw json.RawMessage
-	if raw, ok := proxy["deprecated"]; ok {
-		depRaw = raw
-		delete(proxy, "deprecated")
-	}
-	rest, err := json.Marshal(proxy)
-	if err != nil {
-		return err
-	}
-	type Alias Parameter
-	if err := json.Unmarshal(rest, (*Alias)(p)); err != nil {
-		return err
-	}
-	if len(depRaw) > 0 && string(depRaw) != "null" {
-		var dep Deprecated
-		if !decodeDeprecatedField(&dep, depRaw) {
-			return fmt.Errorf("invalid type for deprecated field")
-		}
-		p.Deprecated = dep
-	}
-	return nil
-}
-
-func (v *VariableDeclaration) UnmarshalJSON(data []byte) error {
-	var proxy map[string]json.RawMessage
-	if err := json.Unmarshal(data, &proxy); err != nil {
-		return err
-	}
-	var depRaw json.RawMessage
-	if raw, ok := proxy["deprecated"]; ok {
-		depRaw = raw
-		delete(proxy, "deprecated")
-	}
-	rest, err := json.Marshal(proxy)
-	if err != nil {
-		return err
-	}
-	type Alias VariableDeclaration
-	if err := json.Unmarshal(rest, (*Alias)(v)); err != nil {
-		return err
-	}
-	if len(depRaw) > 0 && string(depRaw) != "null" {
-		var dep Deprecated
-		if !decodeDeprecatedField(&dep, depRaw) {
-			return fmt.Errorf("invalid type for deprecated field")
-		}
-		v.Deprecated = dep
-	}
-	return nil
-}
-
 func decodeDeprecatedField(dst *Deprecated, data json.RawMessage) bool {
 	if dst == nil {
 		return true
